refactored time-lagged COV from GEC
